+++ packages/expo-web-browser/android/src/main/java/expo/modules/webbrowser/error/PackageManagerNotFoundException.kt
@@ -1,5 +1,0 @@
-package expo.modules.webbrowser.error
-
-import expo.modules.core.errors.CodedException
-
-class PackageManagerNotFoundException : CodedException("Package Manager not found!")